test: PlayerName 테스트 분리
