+++ src/test/java/domain/PlayerTest.java
@@ -1,15 +1,11 @@
 package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatNoException;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import domain.user.Player;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 public class PlayerTest {
 
@@ -18,40 +14,6 @@ public class PlayerTest {
     @BeforeEach
     void setPlayer() {
         player = new Player("kiara");
-    }
-
-    @DisplayName("이름은 5글자 이하이다")
-    @ParameterizedTest
-    @ValueSource(strings = {"a", "ash", "kiara", "woowa"})
-    void validateNameLength(String name) {
-        assertThatNoException().isThrownBy(() -> new Player(name));
-    }
-
-    @DisplayName("이름이 5글자 초과면 예외가 발생한다")
-    @ParameterizedTest
-    @ValueSource(strings = {"hijava", "helloworld", "woowacourse"})
-    void nameNotOver5(String name) {
-        assertThatThrownBy(() -> new Player(name))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessageContaining("플레이어의 이름은 ")
-            .hasMessageContaining("보다 길 수 없습니다.");
-    }
-
-    @DisplayName("이름이 공백이면 예외가 발생한다")
-    @ParameterizedTest
-    @ValueSource(strings = {"", " "})
-    void nameNotEmpty(String name) {
-        assertThatThrownBy(() -> new Player(name))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("플레이어의 이름은 공백으로만 이루어질 수 없습니다.");
-    }
-
-    @DisplayName("이름이 딜러이면 예외가 발생한다")
-    @Test
-    void nameNot딜러() {
-        assertThatThrownBy(() -> new Player("딜러"))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("딜러라는 이름은 사용할 수 없습니다.");
     }
 
     @DisplayName("카드 점수가 21보다 작으면 카드를 받을 수 있다")
